Add CopyIcon and update copyToClipboard function
A new file, CopyIcon.svelte, has been added representing a new SVG icon. The copyToClipboard function in +page.svelte has been updated to receive copy_content as a parameter. This allows different text contents to be copied to the clipboard, improved with the addition of CopyIcon buttons for source and result text